+++ CHANGELOG.md
@@ -73,6 +73,7 @@
   * Fixed dependency of library **Xerces-c** on package
   * Fixed minor typo in the simulation data section of the documentation
   * Fixed the `config.py` to read the `.osm ` files in proper `utf-8` encoding
+  * Created `mayavi_lidar.py` PythonAPI example to display lidar point clouds in real time as an alternative to Open3D.
 
 ## CARLA 0.9.10